Update requirements and move module to system
